Fixes from upstream Selenium2PHPUnit

+++ build/Selenium2PHPUnit/Selenium2TestCase.php
@@ -633,51 +633,15 @@ abstract class Selenium2TestCase extends PHPUnit_Framework_TestCase
 	{
 		if (!empty($this->coverageScriptUrl)) {
 			$coverage = array();
-			foreach (glob($this->coverageScriptUrl."/*".$this->testId) as $file) {
+			$files = glob($this->coverageScriptUrl."/*".$this->testId);
+			if (count($files) == 1) {
+				$file = $files[0];
 				$data = unserialize(file_get_contents($file));
 				unlink($file);
 				unset($file);
-				$filter = new PHP_CodeCoverage_Filter();
-				 /*
-				foreach ($data as $file => $lines) {
-					if ($filter->isFile($file)) {
-						if (!isset($coverage[$file])) {
-							$coverage[$file] = array(
-									'md5' => md5_file($file), 'coverage' => $lines
-							);
-						} else {
-							foreach ($lines as $line => $flag) {
-								if (!isset($coverage[$file]['coverage'][$line]) ||
-										$flag > $coverage[$file]['coverage'][$line]) {
-									$coverage[$file]['coverage'][$line] = $flag;
-								}
-							}
-						}
-					}
-				}
-				*/
-			}
-			
-			return $data;
-		}
-		/*
-		$url = sprintf(
-				'%s?PHPUNIT_SELENIUM_TEST_ID=%s',
-				$this->coverageScriptUrl,
-				$this->testId
-		);
-
-		$buffer = @file_get_contents($url);
-
-		if ($buffer !== FALSE) {
-			$coverageData = unserialize($buffer);
-			if (is_array($coverageData)) {
-				return $this->matchLocalAndRemotePaths($coverageData);
-			} else {
-				throw new Exception('Empty or invalid code coverage data received from url "' . $url . '"');
+				return $data;
 			}
 		}
-		*/
 	
 
 	return array();
